Update drive_lcd.py
Add Home, Clear, text. and cursor positioning functions.

It still doesn't work reliably, but I don;t know if this is
(a) A dodgy physical connection.
(b) A dodgy reading of the GPIO pinout and usage for the Pi 2.
(c) A software problem.

I'm leaning away from (c) because I've done this before and the timings
are stupidly long rather than too tight.